♻ :: profile 훅 분리
