+++ src/Components/Profile/Profile.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React from 'react';
 import * as S from '@/Components/Profile/profile.style';
 import Correction from '@/Components/Profile/Correction/Correction';
 import SettingProfile from '@/Components/Profile/SettingProfile/SettingProfile';
@@ -9,101 +9,20 @@ import CorrectionImg from '@/assets/image/profile/CorrectionImg.svg';
 import ProfileDivider from '@/assets/image/profile/ProflieDivider.svg';
 import Divider from '@/assets/image/profile/Divider.svg';
 
-import { SeugiCustomAxios } from '@/Api/SeugiCutomAxios';
+import useProfile from '@/Hooks/Profile/index';
 
 const Profile = () => {
-    const workspaceId = window.localStorage.getItem("workspaceId");
-    const [isEditing, setIsEditing] = useState(null);
-    const [isSettingOpen, setIsSettingOpen] = useState(false);
-    const [name, setName] = useState('');
-    const [profileData, setProfileData] = useState({
-        status: "",
-        spot: "",
-        belong: "",
-        phone: "",
-        wire: "",
-        location: ""
-    });
-    const dialogRef = useRef<HTMLDivElement>(null);
-
-    useEffect(() => {
-        const fetchProfileData = async () => {
-            try {
-                const res = await SeugiCustomAxios.get(`/profile/me?workspaceId=${workspaceId}`);
-                console.log(res.data);
-
-                setProfileData(res.data.data);
-                setName(res.data.data.member.name);
-            } catch (error) {
-                console.error('Failed to fetch profile data.', error);
-            }
-        };
-        fetchProfileData();
-    }, [workspaceId]);
-
-    const startEditing = (field: any) => {
-        setIsEditing(field);
-    };
-
-    const saveProfileData = async (field: any, value: string) => {
-        try {
-            await SeugiCustomAxios.patch(`/profile/${workspaceId}`,
-                {
-                    status: field === "status" ? value : profileData.status,
-                    spot: field === "spot" ? value : profileData.spot,
-                    belong: field === "belong" ? value : profileData.belong,
-                    phone: field === "phone" ? value : profileData.phone,
-                    wire: field === "wire" ? value : profileData.wire,
-                    location: field === "location" ? value : profileData.location,
-                },);
-
-            setProfileData(prevData => ({
-                ...prevData,
-                [field]: value
-            }));
-
-            setIsEditing(null);
-        } catch (error) {
-            console.error('프로필 저장 실패', error);
-        }
-    };
-
-    const cancelEditing = () => {
-        setIsEditing(null);
-    };
-
-    const toggleSetting = () => {
-        setIsSettingOpen(prevState => !prevState);
-    };
-
-    useEffect(() => {
-        const handleClickOutside = (e: MouseEvent) => {
-            const target = e.target as Node | null;
-
-            if (
-                dialogRef.current &&
-                !dialogRef.current.contains(target) &&
-                !(target && (target as Element).closest('.SettingButton'))
-            ) {
-                setIsSettingOpen(false);
-            }
-        };
-
-        if (isSettingOpen) {
-            document.addEventListener("mousedown", handleClickOutside);
-        }
-
-        return () => {
-            document.removeEventListener("mousedown", handleClickOutside);
-        };
-    }, [isSettingOpen]);
-
-
-
-    const handleNameChange = (newName: string) => {
-        setName(newName);
-    };
-
+    const {
+        name,
+        isEditing,
+        profileData,
+        isSettingOpen,
+        startEditing,
+        saveProfileData,
+        cancelEditing,
+        toggleSetting,
+        handleNameChange
+    } = useProfile();
     return (
         <>
             <S.MyProfileDialog>
